birdr: Initial revision
Initially, the 'init' command is added to initialize the species (and
categories) database tables from a given ebird.org taxonomy list; the
ebird list is optional. Initialization is not yet re-entrant and will
throw an error if it fails.

Initially, I've divided this into model, view, and controller. My
intention is to have view handle command-line parsing and TUI code,
model to hold the object-oriented ORM database accesses, and
controller to act as a procedural glue layer between the two.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>