setup: Add script for initializing a virtualenv
While I prefer guix for setting up my dev environment, it'll be useful
to have a venv setup as well. The setup script will create the venv
and download/upgrade all the required packages (right now, just click
and sqlalchemy). The script is re-entrant so it can be re-run to
upgrade packages.

To run the script:

./pre-inst-env.sh setup

Once the venv exists, it will be activated as part of pre-inst-env.sh
so that binaries and libraries from the venv will be used
automatically.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>