model: Add checklists to the database
A checklist is just a list of species representing some set of birds
that I'd like to see (e.g. "birds of Costa Rica"). A species can be in
multiple checklists, so there's a many-to-many relationship with the
species table. Model this relationship using a bridge table so that
both Checklist and Species can back-populate the other.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>