model: Move the session to a context manager
The new context manager is a Transaction object with methods that
manipulate the database. This way, the session scope is controlled by
the controller (as it ought to be) and the commit can be moved to the
end of the context manager (assuming successful exit).

This has the added benefit of making it easier for the controller to
compose different model operations and work with more complex view
interfaces.

Note that moving the main model methods to the transaction context
means it's impossible to manipulate the database without an active
session.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>