model: Add interface for creating and modifying checklists
The basic operations are "create a new checklist" and "add a species
to a checklist". It is expected that both the species and the
checklist are created (possibly in the same session) before the
species can be added to the database.

There's a slight performance hit to looking up the checklist every
time a new species is added (since usually several species are added
at once). I may change this in the future, but premature optimization
is the root of all evil.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>