view: Add the 'checklist' command
This command takes a single argument that is the checklist name. Then,
it will interactively prompt the user for species names until EOF is
reached.

A new checklist with the given name is created in the database and all
the species are added to the new checklist. This all happens inside a
single transaction, so if an error occurs, the whole operation is
aborted.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>