view: Add the 'show' command
This command takes a checklist name as an argument and will
pretty-print completion data using rich.

For each category of birds in the checklist, a "completion" percentage
is printed indicating the percent of species in the category that have
been observed. Each species is printed alongside a check mark if it
has been observed (or an empty box if not observed).

The overall completion percentage (measured by species observed) is
also printed at the top of the list.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>